(598) Déplacer la logique de création de users dans un service
